fix: rename feature "AccommodationDetail" -> "favourite"
update favoritesSlice import path in store and AccommodationDetail
