+++ UI/src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import authReducer from "../features/auth/authSlice";
-import favoritesReducer from "../features/accommodationDetail/favoritesSlice";
+import favoritesReducer from "../features/favourite/favoritesSlice";
 import bookingReducer from "../features/booking/bookingSlice";
 import searchReducer from "../features/search/searchSlice";
